Realized the Exception code for hourly_pay was using an int as an argument, when it was supposed to use a double. I have changed this in the Production worker .h code.

+++ Factory Worker with Encapsulted Exceptions1/ProductionWorker.h
@@ -39,7 +39,7 @@ public:
 		int value;
 
 	public:
-		InvalidPayrate(int val)
+		InvalidPayrate(double val)
 		{
 			value = val;
 		}
